Added "partial class" functionality
This change allows to extend a chic class with more functions or
properties, without inheriting. Very similar to what a mixin would do,
or what a partial class in C# means. This means, for example, that you
can add funcs/props to a class in separate .js files. This is related to
issue #9 -> https://github.com/rowanmanning/chic/issues/9

A line can be saved by replacing "applySuper" with its value instead,
but it is left as is for better code comprehension.